KDZ_3/payments_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
kafka-python
pydantic
